Split out database code. Add formatting and linting settings. Fix formatting
