+++ .gitignore
@@ -84,9 +84,9 @@ celerybeat-schedule
 
 # Environments
 .env
-.venv
+.venv*
 env/
-venv/
+venv*/
 ENV/
 env.bak/
 venv.bak/
