+++ requirements/main.in
@@ -1,0 +1,7 @@
+Flask>=1.0.2
+graphene>=2.1.8
+gunicorn>=19.8.1
+sqlalchemy
+flask-graphql
+graphene-sqlalchemy
+pytest-runner
