+++ requirements/main_3.7.txt
@@ -1,0 +1,79 @@
+#
+# This file is autogenerated by pip-compile with python 3.7
+# To update, run:
+#
+#    pip-compile --no-emit-index-url --output-file=requirements/main_3.7.txt --pip-args='--no-input' requirements/main.in
+#
+aniso8601==7.0.0
+    # via graphene
+click==8.0.3
+    # via flask
+flask==2.0.2
+    # via
+    #   -r requirements/main.in
+    #   flask-graphql
+flask-graphql==2.0.1
+    # via -r requirements/main.in
+graphene==2.1.9
+    # via
+    #   -r requirements/main.in
+    #   graphene-sqlalchemy
+graphene-sqlalchemy==2.3.0
+    # via -r requirements/main.in
+graphql-core==2.3.2
+    # via
+    #   flask-graphql
+    #   graphene
+    #   graphql-relay
+    #   graphql-server-core
+graphql-relay==2.0.1
+    # via graphene
+graphql-server-core==1.2.0
+    # via flask-graphql
+greenlet==1.1.2
+    # via sqlalchemy
+gunicorn==20.1.0
+    # via -r requirements/main.in
+importlib-metadata==4.8.1
+    # via
+    #   click
+    #   sqlalchemy
+itsdangerous==2.0.1
+    # via flask
+jinja2==3.0.2
+    # via flask
+markupsafe==2.0.1
+    # via jinja2
+promise==2.3
+    # via
+    #   graphene-sqlalchemy
+    #   graphql-core
+    #   graphql-relay
+    #   graphql-server-core
+pytest-runner==5.3.1
+    # via -r requirements/main.in
+rx==1.6.1
+    # via graphql-core
+singledispatch==3.7.0
+    # via graphene-sqlalchemy
+six==1.16.0
+    # via
+    #   graphene
+    #   graphene-sqlalchemy
+    #   graphql-core
+    #   graphql-relay
+    #   promise
+    #   singledispatch
+sqlalchemy==1.4.25
+    # via
+    #   -r requirements/main.in
+    #   graphene-sqlalchemy
+typing-extensions==3.10.0.2
+    # via importlib-metadata
+werkzeug==2.0.2
+    # via flask
+zipp==3.6.0
+    # via importlib-metadata
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
