+++ requirements/test.in
@@ -1,0 +1,3 @@
+pytest>=3.0.7
+flake8
+yapf
