+++ requirements/test_3.7.txt
@@ -1,0 +1,41 @@
+#
+# This file is autogenerated by pip-compile with python 3.7
+# To update, run:
+#
+#    pip-compile --no-emit-index-url --output-file=requirements/test_3.7.txt --pip-args='--no-input' requirements/test.in
+#
+attrs==21.2.0
+    # via pytest
+flake8==4.0.1
+    # via -r requirements/test.in
+importlib-metadata==4.2.0
+    # via
+    #   flake8
+    #   pluggy
+    #   pytest
+iniconfig==1.1.1
+    # via pytest
+mccabe==0.6.1
+    # via flake8
+packaging==21.0
+    # via pytest
+pluggy==1.0.0
+    # via pytest
+py==1.10.0
+    # via pytest
+pycodestyle==2.8.0
+    # via flake8
+pyflakes==2.4.0
+    # via flake8
+pyparsing==2.4.7
+    # via packaging
+pytest==6.2.5
+    # via -r requirements/test.in
+toml==0.10.2
+    # via pytest
+typing-extensions==3.10.0.2
+    # via importlib-metadata
+yapf==0.31.0
+    # via -r requirements/test.in
+zipp==3.6.0
+    # via importlib-metadata
