+++ setup.cfg
@@ -1,29 +1,26 @@
 [metadata]
 name = graphene_flask_example
-version = 0.0.2
+version = 0.1.0
 author = Ayla Khan
 author_email = ayla.khan@gmail.com
 
-[options]
-python_requires >= 3.6
-install_requires =
-    Flask==1.0.2
-    graphene==2.1.8
-    gunicorn==19.8.1
-    SQLAlchemy==1.2.18
-    Flask-GraphQL
-    graphene-sqlalchemy
-    graphql-core
-    graphql-relay
-    graphql-server-core
-    gql
-    pytest-runner
+[flake8]
+max-line-length = 120
+ignore = E122,E123,E124,E126,E127,E128,E261,E262,E265,E722,E731,W391,W503
+exclude = tests, requirements, benchmarks, .git, venv*, example/raw_data
 
-tests_require =
-    pytest>=3.0.7
-    pytest-html
-    pytest-cov
-    teamcity-messages
+[yapf]
+based_on_style = google
+allow_multiline_dictionary_keys = True
+allow_split_before_dict_value = False
+blank_line_before_nested_class_or_def = False
+coalesce_brackets = True
+column_limit = 120
+split_before_named_assigns = False
+split_before_logical_operator = True
+
+[options]
+python_requires >= 3.7
 
 [aliases]
 test=pytest
