+++ setup.py
@@ -1,5 +1,4 @@
 from setuptools import setup
 
-
 if '__main__' == __name__:
-	setup()
+    setup()
