+++ tests/conftest.py
@@ -1,0 +1,21 @@
+import pytest
+import sqlalchemy as sa
+import sqlalchemy.ext.declarative as sa_decl
+import sqlalchemy.orm as sa_orm
+
+
+test_db_url = "sqlite://"
+
+@pytest.fixture(scope="function")
+def session_factory():
+    engine = sa.create_engine(test_db_url)
+    Base = sa_decl.declarative_base()
+    Base.metadata.create_all(engine)
+
+    yield sa_orm.sessionmaker(bind=engine)
+
+    engine.dispose()
+
+@pytest.fixture(scope="function")
+def session(session_factory):
+    return session_factory()
